breaking: rename current checks into leagacy checks
