+++ tests/test_legacy_checks.py
@@ -9,7 +9,7 @@ from multiversx_sdk.network_providers.http_resources import (
 from mxops.data.execution_data import InternalContractData, ScenarioData
 from mxops.errors import CheckFailed
 from mxops.execution.account import AccountsManager
-from mxops.execution.checks import TransfersCheck
+from mxops.execution.legacy_checks import TransfersCheck
 from mxops.execution.msc import ExpectedTransfer, OnChainTransfer
